Safari-611.1.10/LayoutTests/storage/indexeddb/modern/workers-enable-expected.txt

[Worker] Check to make sure we can enable IndexedDB in workers via a runtime setting

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/workers-enable.js
PASS [Worker] self.indexedDB is defined.
PASS [Worker] self.indexedDB is non-null.
PASS [Worker] 'indexedDB' in self is true
PASS [Worker] self.IDBCursor is defined.
PASS [Worker] self.IDBCursor is non-null.
PASS [Worker] 'IDBCursor' in self is true
PASS [Worker] self.IDBCursorWithValue is defined.
PASS [Worker] self.IDBCursorWithValue is non-null.
PASS [Worker] 'IDBCursorWithValue' in self is true
PASS [Worker] self.IDBDatabase is defined.
PASS [Worker] self.IDBDatabase is non-null.
PASS [Worker] 'IDBDatabase' in self is true
PASS [Worker] self.IDBFactory is defined.
PASS [Worker] self.IDBFactory is non-null.
PASS [Worker] 'IDBFactory' in self is true
PASS [Worker] self.IDBIndex is defined.
PASS [Worker] self.IDBIndex is non-null.
PASS [Worker] 'IDBIndex' in self is true
PASS [Worker] self.IDBKeyRange is defined.
PASS [Worker] self.IDBKeyRange is non-null.
PASS [Worker] 'IDBKeyRange' in self is true
PASS [Worker] self.IDBObjectStore is defined.
PASS [Worker] self.IDBObjectStore is non-null.
PASS [Worker] 'IDBObjectStore' in self is true
PASS [Worker] self.IDBOpenDBRequest is defined.
PASS [Worker] self.IDBOpenDBRequest is non-null.
PASS [Worker] 'IDBOpenDBRequest' in self is true
PASS [Worker] self.IDBRequest is defined.
PASS [Worker] self.IDBRequest is non-null.
PASS [Worker] 'IDBRequest' in self is true
PASS [Worker] self.IDBTransaction is defined.
PASS [Worker] self.IDBTransaction is non-null.
PASS [Worker] 'IDBTransaction' in self is true
PASS [Worker] self.IDBVersionChangeEvent is defined.
PASS [Worker] self.IDBVersionChangeEvent is non-null.
PASS [Worker] 'IDBVersionChangeEvent' in self is true
PASS [Worker] self.indexedDB instanceof IDBFactory is true
PASS successfullyParsed is true

TEST COMPLETE
